Cleaning up some code for offline work
