+++ Big Neon/Config/AnalyticsService.swift
@@ -15,17 +15,7 @@ public enum ErrorType: String {
 
 public class AnalyticsService {
 
-    public class var shared: AnalyticsService {
-        struct Static {
-            static let instance: AnalyticsService = AnalyticsService()
-        }
-        return Static.instance
-    }
-
-    public init() {
-    }
-
-    public func reportError(errorType: ErrorType, error: String) {
+    class func reportError(errorType: ErrorType, error: String) {
         let errorAtributes = [errorType.rawValue: error]
         Answers.logCustomEvent(withName: "Error Reported", customAttributes: errorAtributes)
     }
